Initial camera preview
This is the initial camera preview. The wikitude sdk is implemented and
things such as the manifest requirements and permissions have been
modified to fit the final app. Next UI stage will look into adding the
Tool bar and drawer as well as handling orientation change without
having to reload wikitude.